Update the enclaveos-signer to work with SGX2 signatures.
EnclaveOS now supports running enclaves with smaller memory
overhead when run on systems that support SGX2. To take advantage
of lower memory usage, the initial enclave will be constructed
with less memory initially added to the enclave. This results
in a different enclave hash, which requires a separate SGX
signature. So to take advantage of lower memory usage when
SGX2 is present, but still be able to run on hardware that
only supports SGX1, EnclaveOS now signs containers with two
signatures, so the best memory layout can be automatically used
based on what hardware is present.

The enclaveos-signer tool has been updated to allow re-signing
both signatures with your key, or optionally re-sign just
the SGX1 or just the SGX2 layout. More details are in the
README.md file.

Update the requirements.txt file for the new dependencies for
the new version of enclaveos-signer.